+++ FightingGameProject/iPlayer.h
@@ -16,7 +16,7 @@ namespace RB::Players
 		virtual PlayerID GetPlayerID() = 0;
 		virtual iPlayer* GetOtherPlayer() = 0;
 		virtual bool OtherPlayerIsOnRightSide() = 0;
-		virtual int GetStateMachineID() = 0;
+		virtual size_t GetStateMachineID() = 0;
 		virtual void Move(olc::vi2d moveAmount) = 0;
 		virtual olc::vi2d GetPosition() = 0;
 		virtual void SetPosition(olc::vi2d pos) = 0;
