+++ FightingGameProject/iStateMachine.h
@@ -15,8 +15,8 @@ namespace RB::States
 
 	public:
 		virtual void QueueNextState(iState* state) = 0;
-		virtual void SetID(int ID) = 0;
-		virtual int GetID() = 0;
+		virtual void SetID(size_t ID) = 0;
+		virtual size_t GetID() = 0;
 		virtual iState* GetCurrentState() = 0;
 		virtual bool IsTransitioning() = 0;
